+++ Bot_NetCore/Entities/Inviter.cs
