+++ Bot_NetCore/Entities/InviterList.cs
