+++ Bot_NetCore/Examples/inviters.example.xml
