--- SeaOfThieves_Rework/settings.example.xml
@@ -1,65 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Settings>
-    <Version>2.6.0</Version>
-    <Guild>406840993535492109</Guild>
-    <Token>top secret token</Token>
-    <Prefix>!</Prefix>
-    
-    <BotRole>435393449605464074</BotRole>
-    
-    <OkEmoji>✅</OkEmoji>
-    <ErrorEmoji>⛔</ErrorEmoji>
-    
-    <RolesChannel>474870448157491200</RolesChannel>
-    
-    <AutocreateCategory>435427887621996548</AutocreateCategory>
-    <AutocreateGalleon>533876650371186698</AutocreateGalleon>
-    <AutocreateBrigantine>000000000000000000</AutocreateBrigantine>
-    <AutocreateSloop>111111111111111111</AutocreateSloop>
-    <AutocreateSymbol>⛵</AutocreateSymbol>
-    <Bitrate>48000</Bitrate>
-    <FastCooldown>30</FastCooldown>
-    <WaitingRoom>496614051728588804</WaitingRoom>
-
-    <FleetCategory>699687487743721503</FleetCategory>
-    <FleetChillChannel>699688642238611531</FleetChillChannel>
-    <FleetLobby>699688711226392657</FleetLobby>
-    <FleetUserLimiter>5</FleetUserLimiter>
-    
-    <ShipXML>ships.xml</ShipXML>
-    <PrivateCategory>435448014614822923</PrivateCategory>
-    <PrivateRequestsChannel>459986453972123658</PrivateRequestsChannel>
-    <MaxPrivateShips>3</MaxPrivateShips>
-    
-    <DonatorXML>donators.xml</DonatorXML>
-    <DonatorRole>459649927732133888</DonatorRole>
-    <DonatorChannel>648832699200569345</DonatorChannel>
-    <DonatorMessage>558231268773593098</DonatorMessage>
-    
-    <WarningsXML>warnings.xml</WarningsXML>
-    <ModlogChannel>479674576369090560</ModlogChannel>
-    <FulllogChannel>479674700218499072</FulllogChannel>
-    <UserlogChannel>481316315215429672</UserlogChannel>
-    <ErrorLog>672017814932029461</ErrorLog>
-    
-    <BanXML>bans.xml</BanXML>
-    
-    <IgnoredChannels>
-        479674576369090560,
-        479674700218499072,
-        481316315215429672,
-        480288816297672705,        
-    </IgnoredChannels>
-    
-    <AdminRoles>
-        406849001611067393,
-        406850194983092224,
-        435393449605464074,
-		514282258958385152,
-    </AdminRoles>
-	
-	<Developer>261137595965243393</Developer>
-    
-    <CodexChannel>672056183774773268</CodexChannel>
-    <CodexRole>672418978563620865</CodexRole>
-</Settings>
